logout
intine logout wis

+++ routes/web.php
@@ -4,8 +4,10 @@ use App\Http\Controllers\HomeController;
 use App\Http\Controllers\AboutController;
 use App\Http\Controllers\AdminController;
 use App\Http\Controllers\ContactController;
+use App\Http\Controllers\LoginController;
 use App\Http\Controllers\MasterController;
 use App\Http\Controllers\ProjectsController;
+use App\Http\Controllers\RegistrasiController;
 use App\Http\Controllers\SiswaController;
 use Faker\Guesser\Name;
 use Illuminate\Support\Facades\Route;
@@ -25,34 +27,41 @@ use Illuminate\Support\Facades\Route;
 // Route::get('/halaman2', [HomeController::class, 'halaman2']);
 // Route::get('/portofolio', [HomeController::class, 'portofolio']);
 
-// @quest
+// @guest
 Route::get('/', [HomeController::class, 'index']) ->name('home');
 Route::get('/about', [AboutController::class, 'index']) ->name('about');
 Route::get('/projects', [ProjectsController::class, 'projecthome']) ->name('projecthome');
 Route::get('/contact', [ContactController::class, 'contact']) ->name('contact');
 
 // @auth
-Route::get('/admin', [AdminController::class, 'index']) ->name('admin');
+Route::get('/admin', [AdminController::class, 'index']) ->name('admin')->middleware('auth');
 
 // Master Siswa
-Route::get('/admin/siswa', [SiswaController::class, 'index']) ->name('siswa.index');
-Route::get('/admin/siswa/create', [SiswaController::class, 'create']) ->name('siswa.create');
-Route::post('/admin/siswa/store', [SiswaController::class, 'store']) ->name('siswa.store');
-Route::get('/admin/siswa/{id}/edit', [SiswaController::class, 'edit']) ->name('siswa.edit');
-Route::put('/admin/siswa/{id}', [SiswaController::class, 'update']) ->name('siswa.update');
-Route::delete('/admin/siswa/{id}', [SiswaController::class, 'delete']) ->name('siswa.delete');
+Route::get('/admin/siswa', [SiswaController::class, 'index']) ->name('siswa.index')->middleware('auth');
+Route::get('/admin/siswa/create', [SiswaController::class, 'create']) ->name('siswa.create')->middleware('auth');
+Route::post('/admin/siswa/store', [SiswaController::class, 'store']) ->name('siswa.store')->middleware('auth');
+Route::get('/admin/siswa/{id}/edit', [SiswaController::class, 'edit']) ->name('siswa.edit')->middleware('auth');
+Route::put('/admin/siswa/{id}', [SiswaController::class, 'update']) ->name('siswa.update')->middleware('auth');
+Route::delete('/admin/siswa/{id}', [SiswaController::class, 'delete']) ->name('siswa.delete')->middleware('auth');
 
 // Master Project
-Route::resource('/admin/project', ProjectsController::class);
-Route::get('/admin/project/{id}/create', [ProjectsController::class, 'add'])->name('project.add');
+Route::resource('/admin/project', ProjectsController::class)->middleware('auth');
+Route::get('/admin/project/{id}/create', [ProjectsController::class, 'add'])->name('project.add')->middleware('auth');
 
 // Master Contact
-Route::get('/admin/contact', [ContactController::class, 'index']) ->name('contact.index');
-Route::get('/admin/contact/create', [ContactController::class, 'create']) ->name('contact.create');
-Route::post('/admin/contact/store', [ContactController::class, 'store']) ->name('contact.store');
-Route::get('/admin/contact/{id}/edit', [ContactController::class, 'edit']) ->name('contact.edit');
-Route::put('/admin/contact/{id}', [ContactController::class, 'update']) ->name('contact.update');
-Route::delete('/admin/contact/{id}', [ContactController::class, 'delete']) ->name('contact.delete');
+Route::get('/admin/contact', [ContactController::class, 'index']) ->name('contact.index')->middleware('auth');
+Route::get('/admin/contact/create', [ContactController::class, 'create']) ->name('contact.create')->middleware('auth');
+Route::post('/admin/contact/store', [ContactController::class, 'store']) ->name('contact.store')->middleware('auth');
+Route::get('/admin/contact/{id}/edit', [ContactController::class, 'edit']) ->name('contact.edit')->middleware('auth');
+Route::put('/admin/contact/{id}', [ContactController::class, 'update']) ->name('contact.update')->middleware('auth');
+Route::delete('/admin/contact/{id}', [ContactController::class, 'delete']) ->name('contact.delete')->middleware('auth');
 
+// Login
+Route::get('/login', [LoginController::class, 'index']) ->name('login.index')->middleware('guest');
+Route::post('/login', [LoginController::class, 'authenticate']) ->name('login.auth');
+Route::post('/logout', [LoginController::class, 'logout']) ->name('logout');
 
+// Registrasi
+Route::get('/registrasi', [RegistrasiController::class, 'index']) ->name('registrasi.index');
+Route::post('/registrasi', [RegistrasiController::class, 'store']) ->name('registrasi.store');
